### 2021-10-19 #3831 @whedon
[PRE REVIEW]: Rasusa: Randomly subsample sequencing reads to a specified coverage
**Submitting author:** @mbhall88 (<a href="http://orcid.org/0000-0003-3683-6208">Michael Hall</a>)
**Repository:** <a href="https://github.com/mbhall88/rasusa" target ="_blank">https://github.com/mbhall88/rasusa</a>
**Version:** 0.6.0
