### 2021-10-30 #3831 @mbhall88
Thank you for the feedback @mikldk. I address each comment/question below.

> @mbhall88 - thank you for the submission.
>
> I'll go ahead and invite reviewers. Meanwhile I have a few comments/questions:
>
> For the paper:
>
>     * [ ]  Please write out your affiliation (so the reader knows what "EMBL-EBI" means)

Done.

>
>     * [ ]  Maybe mention massive parallel sequencing? (As opposed to Sanger/single molecule sequencing.)

I am curious as to why? Rasusa is agnostic to how the data were generated.

>
>     * [ ]  "For example, coverage of 5x indicates that each nucleotide in the original genetic material is represented five times in the read file." -- maybe make explicit that this in on average five times?

Done.

>
>
> For the method:
>
>     * [ ]  Say the data is 100x (cf. above, that is an average?); we subsample to 20x using `Rasusa` - as I read it, this is now a minimum, right?

Sorry, I don't quite understand what you are asking? If we subsample to 20x then, theoretically, on average, each position has 20x. Not sure if that is what you are referring to?

>
>     * [ ]  Is it possible to use the quality (for `FASTQ` files) or read length (e.g. prefer longer/shorter reads) in the sampling, or is the sampling performed uniformly amongst reads without considering any of such characteristics?

No. This was a deliberate design choice. As I mention in the paragraph about Filtlong, such filtering means the resulting data is not necessarily representative of the the original data source. See https://github.com/mbhall88/rasusa/issues/16#issuecomment-558168931 for an elaboration of this.

>
>     * [ ]  Is it possible to balance the sampling to take into account the reading direction of the reads (3'->5' / 5'->3')?

No. To do this Rasusa would need to know the strand orientation, which is not represented in FASTA/Q, and it only operates on those file formats.
