### 2021-10-31 #3831 @mikldk
@mbhall88, thanks.

> > For the method:
> > ```
> > * [ ]  Say the data is 100x (cf. above, that is an average?); we subsample to 20x using `Rasusa` - as I read it, this is now a minimum, right?
> > ```
>
> Sorry, I don't quite understand what you are asking? If we subsample to 20x then, theoretically, on average, each position has 20x. Not sure if that is what you are referring to?

You answer my question here: the subsampling to 20x means that the subsample has 20x on average, not as a minimum. That's fine. Is this explictely written in the paper (if not, I would consider doing it, but it is up to you to decide)?
