### 2021-11-04 ecdfd7b
add clarifying statement about output coverage
