### 2021-11-14 #3831 @k3yavi
Hi @mikldk , I am ready with my comments, may I ask ask where can I upload the checklist, as mentioned [here](https://joss.readthedocs.io/en/latest/review_checklist.html).
