### 2021-11-22 #3941 @whedon
[REVIEW]: Rasusa: Randomly subsample sequencing reads to a specified coverage
**Submitting author:** @mbhall88 (<a href="http://orcid.org/0000-0003-3683-6208">Michael Hall</a>)
**Repository:** <a href="https://github.com/mbhall88/rasusa" target ="_blank">https://github.com/mbhall88/rasusa</a>
**Version:** v0.6.1
