### 2021-11-25 9bbd6a4
use correct path to bash inside container [#35]
