### 2021-12-07 #3941 @k3yavi
Hi @mikldk , according to JOSS guidelines, I think the paper and the tool passes all the criteria mentioned in the checklist.

I do have some minor points to mention, is it Ok to mention them here or the review is strictly checklist based ?
