### 2021-12-08 ca421f5
fix entrypoint command
