### 2021-12-13 mbhall88/rasusa#36 @k3yavi
Suggestion: Raise warning when desired level of coverage is not possible
Hi,

This is with respect to https://github.com/openjournals/joss-reviews/issues/3941 .

The tool works perfectly for the designed use case. However I think it'd be great to raise a warning when the desired level of coverage is not possible, for example when the fraction of reads to the size of the genome is not large enough.
