### 2021-12-13 31bb0bf
add warning when requested covg not possible [#36]
