### 2021-12-13 mbhall88/rasusa#36 @mbhall88
Thanks @k3yavi, good suggestion. Done in https://github.com/mbhall88/rasusa/commit/31bb0bfa375ef2f65b27b431901834927e8e7510
