### 2022-01-14 #3941 @holtgrewe
I could not edit the checklist above. I took the one from the joss docs, below are the fields I filled. I need to actually test the software on Monday.

## manuscript

### summary

- actually also useful for benchmarks

### main text

- Can review guideline be addressed? "Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it."
- The link to the software archive appears to not work.

### review considerations

- software license: MIT, license file is there
- scholarly effort: check
- documentation
    - statement of need: check
    - installation instructions: excellent
    - example usage: excellent
    - functionality: TODO
    - tests: excellent
- authorship: check
- novelty: check
- proprietary software: not applicable

### checklist

#### general

- repository URL: OK
- license: OK
- contribution and authorship: OK

#### functionality

- installation: check
- functionality: check
- performance: check

#### documentation

- statement of need: check
- installation: check
- example usage: check
- functionality documentation: check
- automated tests: check
- community guidelines: (1) check, for (2,3): it's a github project with issues, so these work for me - check

#### software paper

- summary: check
- statement of need: check
- state of the field: check
- quality of writing: check
- reference: check
