### 2022-01-20 #3941 @mbhall88
@mikldk I am happy with the final proofs.

Regarding the tagged release: I have Zenodo integrated with the repository already, so when I tag a release, a Zenodo archive will be created automatically. One question before I tag a new release: should I merge the `joss_paper` branch into the master branch and then tag? That way the paper is included with the release?
