### 2022-01-20 #3941 @mikldk
@mbhall88 You don't have to merge the `joss_paper` branch into the `main` branch. It's up to you. Please let me know once you have made a release that is accessible at e.g. Zenodo.
