+++ CHANGELOG.md
@@ -7,10 +7,13 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+## [0.6.1]
+
 ### Added
 
 - Warning if the actual coverage of the file(s) is less than the requested coverage
   [[#36][36]]
+- JOSS manuscript
 
 ### Changed
 
@@ -103,6 +106,7 @@ be 1070.
 [0.4.2]: https://github.com/mbhall88/rasusa/releases/tag/0.4.2
 [0.5.0]: https://github.com/mbhall88/rasusa/releases/tag/0.5.0
 [0.6.0]: https://github.com/mbhall88/rasusa/releases/tag/0.6.0
+[0.6.1]: https://github.com/mbhall88/rasusa/releases/tag/0.6.1
 [19]: https://github.com/mbhall88/rasusa/issues/19
 [22]: https://github.com/mbhall88/rasusa/issues/22
 [27]: https://github.com/mbhall88/rasusa/issues/27
@@ -113,5 +117,5 @@ be 1070.
 [36]: https://github.com/mbhall88/rasusa/issues/36
 [benchmark]: https://github.com/mbhall88/rasusa#benchmark
 [error-blog]: https://nick.groenen.me/posts/rust-error-handling/
-[unreleased]: https://github.com/mbhall88/rasusa/compare/0.6.0...HEAD
+[unreleased]: https://github.com/mbhall88/rasusa/compare/0.6.1...HEAD
 
