+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "rasusa"
 description = "Randomly subsample reads to a specified coverage"
-version = "0.6.0"
+version = "0.6.1"
 authors = ["Michael Hall <michael@mbh.sh>"]
 edition = "2018"
 repository = "https://github.com/mbhall88/rasusa"
