### 2022-01-23 #3941 @mbhall88
Hi @mikldk I have merged `joss_paper` into `master` and tagged a release with v0.6.1. It's on Zenodo at https://doi.org/10.5281/zenodo.5895111
