### 2022-01-24 #3941 @mbhall88
> @mbhall88: Can you add "B." initial to Zenodo? (Or remove it from the paper.)

Changed on Zenodo.
